+++ src/main/java/com/p6FullStack/repository/ThemesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.p6FullStack.model.Themes;
 
 @Repository
-public interface ThemesRepository extends JpaRepository<Themes, Integer> {
+public interface ThemesRepository extends JpaRepository<Themes, Long> {
 
 }
